moved hash functions to its own class, made it static
